Fix-up types for NSLog format string.
git-svn-id: https://svn.code.sf.net/p/chicken/code/trunk/cotvnc@819 9ce1af06-8ae8-4ba5-be20-3b57e42882c5

+++ Source/Session.m
@@ -884,7 +884,7 @@ enum {
     else if (trackingNumber == _bottomTrackingTag)
         _vertScrollFactor = -1;
     else
-        NSLog(@"Unknown trackingNumber %d", trackingNumber);
+        NSLog(@"Unknown trackingNumber %ld", (long)trackingNumber);
 
     if ([self connectionIsFullscreen])
         [self beginFullscreenScrolling];
